pyproject.toml
[tool.poetry]
name = "picture-compressor"
version = "0.1.0"
description = "Console line tool to lossy reduce picture size"
authors = ["NovaHFly <mugiviska@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
compress = "picture_compressor:main_cli"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
cloup = "^3.0.5"

[tool.poetry.group.testing.dependencies]
ruff = "^0.4.10"
mypy = "^1.10.1"
pytest = "^7"
pytest-lazy-fixture = "^0.6.3"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = '--cov picture_compressor'

[tool.mypy]
mypy_path = "picture_compressor"
strict_optional = false
disable_error_code = "import-untyped"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
